requirements.txt
torch>=2.0.0
torchvision
torchaudio
flwr
numpy
pandas
scikit-learn
shap
lime
streamlit
plotly
matplotlib
seaborn
Flask
waitress
requests
PyPDF2
pytesseract
Pillow
